Limit pos. res curve stoic window for pre-processing
When constructing the pos. and neg. electrode resistance curves from the
full cell data, we need to recognize that, at 0% SOC, the positive
electrode lithium stoichiometry is not unity. This means that the
positive electrode resistance curve must necessarily be even larger
below 0% full cell SOC.

This code fix estimates the pos. electrode lithium stoichiometry to be
some value less than 1 to correspond to 0% SOC.